pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "sdx-lc"
version = "3.2.0"
description = "AtlanticWave-SDX project's local controller"
authors = [
    { name = "Yufeng Xin", email = "yxin@renci.org" },
    { name = "Cong Wang", email = "cwang@renci.org" },
    { name = "Sajith Sasidharan", email = "sajith@renci.org" },
    { name = "Italo Valcy", email = "italo@ampath.net" },    
]
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "python_dateutil == 2.8.2",
    "setuptools >= 21.0.0",
    "pika >= 1.2.0",
    "connexion[swagger-ui] == 2.14.2",
    "asgiref >= 3.7.2",    
    "pymongo > 3.0",
    "dotenv",
    "sense-sdx @ git+https://github.com/atlanticwave-sdx/sense-sdx.git",
    "sdx-datamodel @ git+https://github.com/atlanticwave-sdx/datamodel@3.0.0.dev13",
]

[project.optional-dependencies]
test = [
    "flask_testing == 0.8.1",
    "coverage >= 4.0.3",
    "networkx == 2.8.8",
    "pytest >= 7.2.0",
    "pytest-cov >= 4.0.0",
    "pluggy >= 0.3.1",
    "py >= 1.4.31",
    "randomize >= 0.13",
]
format = [
    "black == 24.*",
    "isort == 5.*",
]
wsgi = [
    "uvicorn"
]

[options.packages.find]
where = "sdx_lc"

[tool.black]
include = '\.py?$'

[tool.isort]
profile = "black"
src_paths = ["setup.py", "sdx_lc", "mq-test"]

[tool.pytest.ini_options]
addopts = "--cov=sdx_lc --cov-report term-missing"
testpaths = [
    "sdx_lc/test"
]

[tool.coverage.run]
branch = true
omit = [ "sdx_lc/test/*" ]
relative_files = true

